# src/Archetypical.Software.SchemaRegistry/Controllers/SchemasController.cs
/*
 * Cloud Native Data Schema Registry
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1
 *
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Linq;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Mvc;
using Microsoft.Extensions.Logging;
using Newtonsoft.Json;
using Archetypical.Software.SchemaRegistry.Attributes;
using Archetypical.Software.SchemaRegistry.Shared.Data;
using Archetypical.Software.SchemaRegistry.Shared.Interfaces;
using Microsoft.EntityFrameworkCore;
using Swashbuckle.AspNetCore.Annotations;

namespace Archetypical.Software.SchemaRegistry.Controllers
{
    /// <summary>
    ///
    /// </summary>
    [ApiController]
    public class SchemasController : ControllerBase
    {
        private readonly Context _context;
        private readonly IEnumerable<ISchemaValidator> _validators;
        private readonly ILogger<SchemasController> _logger;

        public SchemasController(Context context, IEnumerable<ISchemaValidator> validators, ILogger<SchemasController> logger)
        {
            _context = context;
            _validators = validators;
            _logger = logger;
        }

        /// <summary>
        /// Delete schema
        /// </summary>
        /// <param name="groupId">schema group</param>
        /// <param name="schemaId">schema id</param>
        /// <response code="204">OK no content</response>
        /// <response code="404">Matching schema not found</response>
        [HttpDelete]
        [Route("/schemagroups/{groupId}/schemas/{schemaId}")]
        [ValidateModelState]
        [SwaggerOperation("DeleteSchema")]
        public virtual async Task<IActionResult> DeleteSchema([FromRoute][Required] string groupId,
            [FromRoute][Required] string schemaId)
        {
            var schemas = await _context.Schemata.Where(x => x.SchemaGroupId == groupId && x.Id == schemaId).ToListAsync();

            if (schemas.Any() == false)
                return NotFound();

            foreach (var schema in schemas)
            {
                _context.Schemata.Remove(schema);
            }

            await _context.SaveChangesAsync();
            return NoContent();
        }

        /// <summary>
        /// Get latest version of schema
        /// </summary>
        /// <remarks>Get latest version of schema.</remarks>
        /// <param name="groupId">schema group</param>
        /// <param name="schemaId">schema id</param>
        /// <response code="200">OK</response>
        [HttpGet]
        [Route("/schemagroups/{groupId}/schemas/{schemaId}")]
        [ValidateModelState]
        [SwaggerOperation("GetLatestSchema")]
        [SwaggerResponse(statusCode: 200, type: typeof(string), description: "OK")]
        public virtual async Task<IActionResult> GetLatestSchema([FromRoute][Required] string groupId,
            [FromRoute][Required] string schemaId)
        {
            var schema = await _context.Schemata.Where(x => x.SchemaGroupId == groupId && x.Id == schemaId)
                .OrderByDescending(x => x.Version).FirstOrDefaultAsync();

            if (schema == null)
                return NotFound();
            var group = await _context.SchemaGroups.FirstAsync(x => x.Id == groupId);
            Response.Headers.Add("Location",
                Url.Action("GetSchemaVersion", "Versions", new { groupId, schemaId, versionnumber = schema.Version }));
            Response.Headers.Add("Schema-Id", schemaId);
            Response.Headers.Add("Schema-Id-Location",
                Url.Action("GetLatestSchema", "Schemas", new { groupId, schemaId }));
            Response.Headers.Add("Schema-Version", schema.Version.ToString());

            return new ContentResult()
            {
                Content = schema.Contents,
                ContentType = _validators.First(x => x.SchemaFormat == group.Format).ContentType,
                StatusCode = 200
            };
        }
    }
}